update: smoothing profile and give nice ui at notes
