+++ src/components/card/CardDefault.tsx
@@ -1,4 +1,7 @@
 import { useId } from "react";
+import data from "@/assets/data.png";
+import laptop from "@/assets/laptop.png";
+import notes from "@/assets/notes.png";
 
 interface CardDefaultProps extends React.HTMLAttributes<HTMLDivElement> {
   title: string;
@@ -7,6 +10,9 @@ interface CardDefaultProps extends React.HTMLAttributes<HTMLDivElement> {
   date?: string;
   className?: string;
 }
+
+export const image = [{ img: data }, { img: laptop }, { img: notes }];
+export const bg = ["#BAB3FF", "#FFD27D"];
 
 export default function CardDefault({
   onClick,
@@ -17,6 +23,7 @@ export default function CardDefault({
   ...props
 }: CardDefaultProps) {
   const id = useId();
+  const titleLength = 100;
 
   const formatDate = (dateStr: string | undefined) => {
     if (!dateStr) return "";
@@ -29,20 +36,40 @@ export default function CardDefault({
     return `${day}-${month}-${year}`;
   };
 
+  const randomImg = image[Math.floor(Math.random() * image.length)];
+  const randomBg = bg[Math.floor(Math.random() * bg.length)];
+
   return (
     <div
-      className={`${className} card p-5 rounded-[20px] bg-[#F7F7F7] hover:bg-[#E8EDFF] h-[250px] flex flex-col justify-between cursor-pointer`}
+      className={`${className} card p-5 rounded-[20px] bg-[#F7F7F7] hover:bg-[#E8EDFF] h-[250px] flex flex-col justify-center items-start cursor-pointer`}
       id={`card-${id}`}
       onClick={onClick}
       {...props}
     >
-      <div className="card-header">
-        <p className="title-text font-semibold text-2xl mt-2 capitalize">{title}</p>
-        <p className="lead font-thin text-lg max-w-xl my-5">{lead}</p>
-      </div>
-      <div className="card-footer">
-        <div className="wrapper-info flex items-center gap-5">
-          <p className="date">{formatDate(date)}</p>
+      <div className="flex items-center gap-5">
+        <div className={`wrapper-img w-4/12 my-auto rounded-[10px] min-w-[150px]`}
+        style={{ backgroundColor: randomBg }}>
+          <img
+            className={` h-[150px] object-cover rounded-10px `}
+            src={randomImg.img}
+          />
+        </div>
+        <div className="wrapper-text w-8/12">
+          <div className="card-header">
+            <p className="title-text font-semibold text-2xl capitalize">
+              {title}
+            </p>
+            <p className="lead font-thin text-lg max-w-xl my-3">
+              {lead.length > titleLength
+                ? lead.slice(0, titleLength) + "..."
+                : lead}
+            </p>
+          </div>
+          <div className="card-footer">
+            <div className="wrapper-info flex items-center gap-5">
+              <p className="date">{formatDate(date)}</p>
+            </div>
+          </div>
         </div>
       </div>
     </div>
