+++ src/router/index.tsx
@@ -5,7 +5,7 @@ import AccessPages from "@/pages/AccessPages";
 import LandingPage from "@/pages/LandingPage";
 import Unauthorized from "@/error/Unauthorized";
 import NotFound from "@/error/NotFound";
-import Profile from "@/pages/Profile";
+import AccessProfile from "@/pages/AccessProfile";
 
 export default function AppRoutes(): JSX.Element {
   return (
@@ -15,7 +15,7 @@ export default function AppRoutes(): JSX.Element {
         <Route path="/unauthorized" element={<Unauthorized />} />
         <Route path="*" element={<NotFound />} />
         <Route element={<PrivateRoute />}>
-          <Route path="profile" element={<Profile />} />
+          <Route path="profile" element={<AccessProfile />} />
 
           <Route path="dashboard" element={<AccessPages />} />
         </Route>
