+++ vite.config.ts
@@ -1,21 +1,21 @@
-import { defineConfig, UserConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-import path from 'path';
+import { defineConfig, UserConfig } from "vite";
+import react from "@vitejs/plugin-react";
+import path from "path";
 
 // https://vite.dev/config/
 const config: UserConfig = defineConfig({
   plugins: [react()],
   resolve: {
     alias: {
-      '@': path.resolve(__dirname, './src'),
+      "@": path.resolve(__dirname, "./src"),
     },
   },
   optimizeDeps: {
-    include: ['react', 'react-dom'],
+    include: ["react", "react-dom"],
   },
   server: {
     mimeTypes: {
-      '.js': 'application/javascript',
+      ".js": "application/javascript",
     },
   },
 });
